release configs and launches
