+++ .gitignore
@@ -2,3 +2,4 @@
 build
 install
 log
+*.pyc
